Move bound checking to function in new file and add printing to log file
